Fix for launch speed
Win - " The spaceBar acts like any other key, so in a text editor you
would get (*space*...*space**space**space**space*) if you held it down.
I added a boolean to the "PressedButton" if statement to fix that. I
also changed it to "CurrentTime" because that always uses the same
reference time. "Nanotime" apparently changes reference points...Idk how
or why lol so I just changed it to be safe "